Added dijkstra

+++ src/main/java/org/buet/sky/airrecalculator/CityMap.java
@@ -43,8 +43,8 @@ public class CityMap {
     @FXML
     private void onAddCity() {
             String name = nameField.getText();
-            double x = Double.parseDouble(xField.getText());
-            double y = Double.parseDouble(yField.getText());
+            int x = Integer.parseInt(xField.getText());
+            int y = Integer.parseInt(yField.getText());
 
 
             City city = new City();
@@ -70,8 +70,8 @@ public class CityMap {
 
         gc.setFill(Color.RED);
         for (City city : cities) {
-            double x = city.getX();
-            double y = city.getY();
+            int x = city.getX();
+            int y = city.getY();
             gc.fillOval(x - 5, y - 5, 10, 10);
             gc.strokeText(city.getName(), x + 7, y);
         }
